Fix race condition when rendering the UI (#304)
This fix also depends on on a fix in ign-gazebo so that it now can call
SwapFromThread, otherwise the race condition will still happen.

At the current moment, the implementation is very wasteful creating many
more RenderTargets than needed (because 2 workspaces are created).
This issue will be fixed in the next commit

Affects ignitionrobotics/ign-rendering#304

Signed-off-by: Matias N. Goldberg <dark_sylinc@yahoo.com.ar>