Explicitly log errors
On incorrect usage we may end up throwing. Developers may be tempted to
"fix" this bug by catching the error and continuing. That would be a
mistake.

This exception is desirable because it indicates there is a bug
somewhere else, so we explicitly rethrow and log it as an ignition
error, asking users who see this to report it.

Signed-off-by: Matias N. Goldberg <dark_sylinc@yahoo.com.ar>