Destroy root node and clean up sensor resources on exit (#617)
When the scene was destroyed, only non-root nodes were destroyed. This PR makes sure to destroy the root node as well. As soon as that is done, I noticed that destructors of sensors are now called but tests are segfaulting. Fixed by cleaning up more resources in sensors' Destroy function.

Signed-off-by: Ian Chen <ichen@osrfoundation.org>