Fix debug builds of gz-rendering-ogre2 plugin (#1014)
The ogre2 plugin fails to load in gz-sim when gz-rendering
is built in Debug mode. The following error message is emitted:
```
Error while loading the library [....lib/gz-rendering-8/engine-plugins/libgz-rendering-ogre2.so]: ....lib/gz-rendering-8/engine-plugins/libgz-rendering-ogre2.so: undefined symbol: _ZThn1008_N4Ogre7HlmsPbs19_changeRenderSystemEPNS_12RenderSystemE
```

This is because we set `DEBUG=1` and `_DEBUG=1`
on the ogre2 target which seem to change the ABI of the generated
library. These defines should really be set if Ogre2 itself is built in
debug mode, so this PR simply comments out the CMake command that adds
those defines.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>