Fix user settings of thermal camera being ignored (#1138)
Fixes #1134

Changes to thermal camera settings that can be made through the plugin
gz-sim-thermal-sensor-system would not show up in the result. While the
value of the data members was updated, the named constants in the
material shader used for rendering were not. We fix that by always
setting the shader constants in PreRender to the current values.

This might be relevant for other thermal camera settings as well, but
this focuses on min and max temperature and resolution, as those can be
modified through the gz-sim-thermal-sensor-system plugin.

---------

Signed-off-by: Marcel Jacobse <44684927+mjacobse@users.noreply.github.com>