Guard against bad memory access in RenderPassSystem (#1153)
Fixes a bad memory access that in RenderPassSystem that happens in some linux builds.

On Ubuntu with gcc, global_##classname##Factory is initialized after RenderPassSystem::renderPassMap, so the Register call in the constructor for ##classname##Factory works as expected.
However, on some linux builds, global_##classname##Factory may be initialized before RenderPassSystem::renderPassMap, leading to a bad memory access.

This is fixed by using the Construct On First Use idiom, where the static renderPassMap member variable is replaced by a GetRenderPassMap() member function in which the actual static render pass map is instantiated. This member function is then called in RenderPassSystem::Register() and in RenderPassSystem::CreateImpl() to write to/ read from the static render map.

Signed-off-by: Shameek Ganguly <shameek@intrinsic.ai>